backend-django/requirements.txt
channels==3.*
channels-redis==3.*
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
django-cors-headers
django-ses
requests
openai
langchain
opencv-python
